+++ src/desktop/Awb.Core/Project/Clients/Esp32ClientHardwareConfig.cs
@@ -9,7 +9,7 @@ using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
 using System.Text.Json.Serialization;
 
-namespace Awb.Core.Project.Various
+namespace Awb.Core.Project.Clients
 {
     public class Esp32ClientHardwareConfig : IProjectObjectListable
     {
@@ -90,28 +90,28 @@ namespace Awb.Core.Project.Various
                     Message = "SSD1306 I2C address has to be set",
                     Category = ProjectProblem.Categories.Various,
                     ProblemType = ProjectProblem.ProblemTypes.Error,
-                    Source = this.TitleShort
+                    Source = TitleShort
                 };
                 if (string.IsNullOrWhiteSpace(Ssd1306ComPins)) yield return new ProjectProblem
                 {
                     Message = "SSD1306 COM pins have to be set",
                     Category = ProjectProblem.Categories.Various,
                     ProblemType = ProjectProblem.ProblemTypes.Error,
-                    Source = this.TitleShort
+                    Source = TitleShort
                 };
                 if (Ssd1306ScreenHeight == null) yield return new ProjectProblem
                 {
                     Message = "SSD1306 screen height has to be set",
                     Category = ProjectProblem.Categories.Various,
                     ProblemType = ProjectProblem.ProblemTypes.Error,
-                    Source = this.TitleShort
+                    Source = TitleShort
                 };
                 if (Ssd1306ScreenWidth == null) yield return new ProjectProblem
                 {
                     Message = "SSD1306 screen width has to be set",
                     Category = ProjectProblem.Categories.Various,
                     ProblemType = ProjectProblem.ProblemTypes.Error,
-                    Source = this.TitleShort
+                    Source = TitleShort
                 };
             }
             yield break;
